requirements.txt
pyyaml
pytest
flake8
black
superfakefalsopaquetesas4d64